% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pandemic.R
\name{plot_pandemic}
\alias{plot_pandemic}
\title{Plot a simulated pandemic}
\usage{
plot_pandemic(
  pan,
  t = NULL,
  color = c("black", "firebrick1", "snow3", "cornflowerblue"),
  sleeptime = 0.1
)
}
\arguments{
\item{pan}{A pandemic object generated by \code{sim_pandemic()}}

\item{t}{Optional: A specific timepoint to plot. If not specified the entire pandemic is plotted over time.}

\item{color}{A color vector corresponding to Susceptible, Infected, Dead, Recovered.}

\item{sleeptime}{A time value for the sleep timer for dynamic plotting of the pandemic (do not go below 0.1).}
}
\description{
With this function you can visualize the pandemics you simulated with \code{sim_pandemic()}.
}
\examples{

# first simulate a pandemic
pandemic = sim_pandemic()

# then plot it
plot_pandemic(pandemic)

# you can also plot individual time points
plot_pandemic(pandemic, t = 200)
}
