% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pandemic.R
\name{sim_pandemic}
\alias{sim_pandemic}
\title{Simulate a pandemic}
\usage{
sim_pandemic(
  n = 500,
  K = 500,
  inf0 = 1,
  mobility = 0.3,
  radius = 0.2,
  inf_prob = 0.7,
  recprob = 0.1,
  resusprob = 0.05,
  deadfactor = 0.005,
  measures = 0.8,
  mobility_reduction = 0.26,
  inf_prob_reduction = 0.4
)
}
\arguments{
\item{n}{Number of time points to simulate}

\item{K}{Number of individuals living in the box}

\item{inf0}{Number of infected individuals at t = 0}

\item{mobility}{A value determining the mobility of the population}

\item{radius}{A value determining the radius in which infections can happen}

\item{inf_prob}{Probability that an individual gets infected when in proximity of an infected individual}

\item{recprob}{Probability to recover if infected (corresponds to a geometric distribution for recover time)}

\item{resusprob}{Probability to become susceptible again after infection and recovery (corresponds to a geometric distribution for the time to become susceptible again)}

\item{deadfactor}{Expected proportion of infected individuals that die}

\item{measures}{Proportion of infected population at which to start countermeasures against the pandemic}

\item{mobility_reduction}{Mobility reduction in percent if measures are active}

\item{inf_prob_reduction}{Reduction of the infection probability in percent if measures are active (proxy for e.g. wearing masks)}
}
\value{
Returns a pandemic object that can be used for plotting the pandemic with the \code{plot_pandemic()} function.
}
\description{
With this function you can simulate a pandemic for a population of dots living in a small box.
For simplicitely they move completly random. You have a lot of options to change the pandemic.
For example you can change the number of individuals \code{K}or increase the population's \code{mobility}. You can also make the disease more infectious or more deadly with \code{inf_prob} and \code{deadfactor}.
You can change how fast individuals recover with \code{recprob} and you can even simulate the effect of countermeasures with \code{measures}, \code{mobility_reduction} and \code{inf_prob_reduction}.
}
\examples{

## Example 1: Simulating a pandemic with default values

pandemic = sim_pandemic()

## Example 2: Changing values
# more aggressive pandemic with stronger countermeasures
pandemic = sim_pandemic(mobility = 0.5, deadfactor = 0.01, measures = 0.8, mobility_reduction = 0.5)
}
